+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.app_AppLanguageLayoutPaparazziTest_AppLanguageLayout()[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
