+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.app_AppLanguageLayoutPaparazziTest_AppLanguageLayout()[Pixel_6_Pro,NOTNIGHT,NO_ACCESSIBILITY].png
