+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.app_AppLanguageLayoutPaparazziTest_AppLanguageLayout()_-_Searching[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
