+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.app_AppLanguageLayoutPaparazziTest_AppLanguageLayout()_-_Searching[Pixel_6_Pro,NOTNIGHT,NO_ACCESSIBILITY].png
